requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
django-ckeditor==5.6.1
django-js-asset==1.1.0
django-markdown==0.8.4
django-mdeditor==0.1.6
idna==2.7
lxml==4.2.3
Markdown==2.6.11
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
